Simplify accumulative hook
